+++ lib/easy_dart_logger.dart
@@ -5,3 +5,4 @@ library easy_dart_logger;
 export 'src/dart_logger_configuration.dart'
     show DartLoggerConfiguration, LogFormat;
 export 'src/easy_dart_logger.dart' show DartLogger, DartLoggerLevel;
+export 'src/no_easy_dart_logger.dart' show NoDartLogger;