+++ CHANGELOG.md
@@ -5,3 +5,11 @@
 ## 1.0.0+2
 
 - Fix README.md
+
+## 1.0.0+3
+
+- Remove some ignored files
+
+## 1.0.1+1
+
+- Add NoDartLogger to override DartLogger with no print